% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admininfo.R
\name{admininfo}
\alias{admininfo}
\title{Get admin information}
\usage{
admininfo(agg.pop, SpatialPolygons, proportion, admin)
}
\arguments{
\item{agg.pop}{aggregated poplulation from aggPopulation function.}

\item{SpatialPolygons}{spatial polygons dataframe for admin 1 or admin 2.}

\item{proportion}{urban/rural proportion from proportion function}

\item{admin}{admin level}
}
\value{
This function returns the 1. dataframe that contains admin 1 and admin 2 information and coordinates for each cluster and 2. Adjacency matrix
\itemize{
  \item admin.info
}
}
\description{
This function get admin information including name, character, population and unban/rural proportion.
}
\examples{
\dontrun{
}

}
\author{
Qianyu Dong
}
