% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggPopulation.R
\name{aggPopulation}
\alias{aggPopulation}
\title{Get population information}
\usage{
aggPopulation(tiff_100m, SpatialPolygons, varname)
}
\arguments{
\item{SpatialPolygons}{spatial polygons dataframe.}

\item{varname}{column name of district name in SpatialPolygons file.}

\item{tiff100}{spatial point dataframe}
}
\value{
This function returns the dataset that contain district name and population for given  tiff files and polygons of admin level
  \item {agg.pop
}
}
\description{
This function aggregate population to particular admin levels
}
\examples{
\dontrun{
}

}
\author{
Qianyu Dong
}
