% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AH_TOBC_W_OTH.R
\name{AH_TOBC_W_OTH}
\alias{AH_TOBC_W_OTH}
\title{AH_TOBC_W_OTH}
\usage{
AH_TOBC_W_OTH(IRdata)
}
\arguments{
\item{IRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
AH_TOBC_W_OTH
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "AH_TOBC_W_OTH", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
