% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AH_TOBU_M_ASM.R
\name{AH_TOBU_M_ASM}
\alias{AH_TOBU_M_ASM}
\title{AH_TOBU_M_ASM}
\usage{
AH_TOBU_M_ASM(MRdata)
}
\arguments{
\item{MRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
AH_TOBU_M_ASM
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
