% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AH_TOBU_W_SNM.R
\name{AH_TOBU_W_SNM}
\alias{AH_TOBU_W_SNM}
\title{AH_TOBU_W_SNM}
\usage{
AH_TOBU_W_SNM(IRdata)
}
\arguments{
\item{IRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
AH_TOBU_W_SNM
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
