% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AH_TOBU_W_SNN.R
\name{AH_TOBU_W_SNN}
\alias{AH_TOBU_W_SNN}
\title{AH_TOBU_W_SNN}
\usage{
AH_TOBU_W_SNN(IRdata)
}
\arguments{
\item{IRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
AH_TOBU_W_SNN
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "AH_TOBU_W_SNN", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
