% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{AN_ANEM_W_ANY}
\alias{AN_ANEM_W_ANY}
\title{Internal function to compute "Any anemia - women"}
\usage{
AN_ANEM_W_ANY(RData)
}
\arguments{
\item{RData}{list returned from getDHSdata when indicator is specified as NULL}
}
\description{
Internal function to compute "Any anemia - women"
}
