% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AN_NUTS_W_OVW.R
\name{AN_NUTS_W_OVW}
\alias{AN_NUTS_W_OVW}
\title{AN_NUTS_W_OVW Women who are overweight according to BMI (25.0-29.9)
nt_wm_ovwt in github
IR}
\usage{
AN_NUTS_W_OVW(Rdata)
}
\arguments{
\item{Rdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
AN_NUTS_W_OVW Women who are overweight according to BMI (25.0-29.9)
nt_wm_ovwt in github
IR
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
