% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AN_NUTS_W_THN.R
\name{AN_NUTS_W_THN}
\alias{AN_NUTS_W_THN}
\title{AN_NUTS_W_THN
IRdata
Women who are thin according to BMI (<18.5)}
\usage{
AN_NUTS_W_THN(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
AN_NUTS_W_THN
IRdata
Women who are thin according to BMI (<18.5)
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "AN_NUTS_W_THN",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::AN_NUTS_W_THN)
}
}
\author{
Qianyu Dong
}
