% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{ch_allvac_either}
\alias{ch_allvac_either}
\title{CH_VACS_C_BAS
Children with all 8 basic vaccinations (age 12-23)
"All basic vaccinations according to either source"}
\usage{
ch_allvac_either(Rdata)
}
\arguments{
\item{Rdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
CH_VACS_C_BAS
Children with all 8 basic vaccinations (age 12-23)
"All basic vaccinations according to either source"
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "CH_VACS_C_BAS",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::ch_allvac_either)
}
}
\author{
Qianyu Dong
}
