% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH_ARIS_C_ADV.R
\name{CH_ARIS_C_ADV}
\alias{CH_ARIS_C_ADV}
\title{CH_ARIS_C_ADV Children with ARI for whom advice or treatment was sought;
ch_ari_care in github
KR}
\usage{
CH_ARIS_C_ADV(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
CH_ARIS_C_ADV Children with ARI for whom advice or treatment was sought;
ch_ari_care in github
KR
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "CH_ARIS_C_ADV", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
