% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{ch_diar_ors_rhf}
\alias{ch_diar_ors_rhf}
\title{CH_DIAT_C_ORT
KR  Diarrhea treatment (Children under five with diarrhea treated with either ORS or RHF)}
\usage{
ch_diar_ors_rhf(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
CH_DIAT_C_ORT
KR  Diarrhea treatment (Children under five with diarrhea treated with either ORS or RHF)
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "CH_DIAT_C_ORT",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::ch_diar_ors_rhf)
}
}
\author{
Qianyu Dong
}
