% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH_DIAT_C_ADV.R
\name{CH_DIAT_C_ADV}
\alias{CH_DIAT_C_ADV}
\title{CH_DIAT_C_ADV Treatment of diarrhea: Advice or treatment was sought;
ch_diar_care in github
KR}
\usage{
CH_DIAT_C_ADV(Rdata)
}
\arguments{
\item{Rdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
CH_DIAT_C_ADV Treatment of diarrhea: Advice or treatment was sought;
ch_diar_care in github
KR
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
