% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH_DIAT_C_AMO.R
\name{CH_DIAT_C_AMO}
\alias{CH_DIAT_C_AMO}
\title{CH_DIAT_C_AMO}
\usage{
CH_DIAT_C_AMO(KRdata)
}
\arguments{
\item{KRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
CH_DIAT_C_AMO
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
