% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH_DIAT_C_ORS.R
\name{CH_DIAT_C_ORS}
\alias{CH_DIAT_C_ORS}
\title{CH_DIAT_C_ORS}
\usage{
CH_DIAT_C_ORS(KRdata)
}
\arguments{
\item{KRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
CH_DIAT_C_ORS
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
