% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH_DIFP_C_FAL.R
\name{CH_DIFP_C_FAL}
\alias{CH_DIFP_C_FAL}
\title{CH_DIFP_C_FAL}
\usage{
CH_DIFP_C_FAL(KRdata)
}
\arguments{
\item{KRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
CH_DIFP_C_FAL
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
