% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH_FEVR_C_FEV.R
\name{CH_FEVR_C_FEV}
\alias{CH_FEVR_C_FEV}
\title{CH_FEVR_C_FEV}
\usage{
CH_FEVR_C_FEV(KRdata)
}
\arguments{
\item{KRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
CH_FEVR_C_FEV
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "CH_FEVR_C_FEV", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
