% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH_FEVT_C_ADV.R
\name{CH_FEVT_C_ADV}
\alias{CH_FEVT_C_ADV}
\title{CH_FEVT_C_ADV Children with fever for whom advice or treatment was sought
ch_fev_care in github ml_fev_care should produce the same data
KR}
\usage{
CH_FEVT_C_ADV(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
CH_FEVT_C_ADV Children with fever for whom advice or treatment was sought
ch_fev_care in github ml_fev_care should produce the same data
KR
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
