% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{ch_meas_either}
\alias{ch_meas_either}
\title{CH_VACC_C_MSL
MCV: Measles
Measles vaccination received
Percentage of children (age 12-23)
ch_meas_either   CH_VAC.do  KR
"Measles vaccination according to either source"}
\usage{
ch_meas_either(Rdata)
}
\arguments{
\item{Rdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
CH_VACC_C_MSL
MCV: Measles
Measles vaccination received
Percentage of children (age 12-23)
ch_meas_either   CH_VAC.do  KR
"Measles vaccination according to either source"
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "CH_VACC_C_MSL",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::ch_meas_either)
}
}
\author{
Qianyu Dong
}
