% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{ch_pent1_either}
\alias{ch_pent1_either}
\title{CH_VACC_C_DP1
KR
Percentage of children (age 12-23)
Pentavalent 1rd dose vaccination according to either source"}
\usage{
ch_pent1_either(Rdata)
}
\arguments{
\item{Rdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
CH_VACC_C_DP1
KR
Percentage of children (age 12-23)
Pentavalent 1rd dose vaccination according to either source"
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "CH_VACC_C_DP1",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::ch_pent1_either)
}
}
\author{
Qianyu Dong
}
