% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH_VACC_C_BAS.R
\name{CH_VACC_C_BAS}
\alias{CH_VACC_C_BAS}
\title{CH_VACC_C_BAS
KRdata
Children with all 8 basic vaccinations (age 12-23)}
\usage{
CH_VACC_C_BAS(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
CH_VACC_C_BAS
KRdata
Children with all 8 basic vaccinations (age 12-23)
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "CH_VACC_C_BAS",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::CH_VACC_C_BAS)
}
}
\author{
Qianyu Dong
}
