% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH_VACC_C_BCG.R
\name{CH_VACC_C_BCG}
\alias{CH_VACC_C_BCG}
\title{CH_VACC_C_BCG  Percentage of children 12-23 months who had received BCG vaccination
ms_afm_15 in github
KR}
\usage{
CH_VACC_C_BCG(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
CH_VACC_C_BCG  Percentage of children 12-23 months who had received BCG vaccination
ms_afm_15 in github
KR
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "CH_VACC_C_BCG", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
