% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH_VACC_C_DP1.R
\name{CH_VACC_C_DP1}
\alias{CH_VACC_C_DP1}
\title{CH_VACC_C_DP1
KRdata
Pentavalent 1st dose vaccination
Percentage of children (age 12-23) with 1st dose of Pentavalent vaccine}
\usage{
CH_VACC_C_DP1(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
CH_VACC_C_DP1
KRdata
Pentavalent 1st dose vaccination
Percentage of children (age 12-23) with 1st dose of Pentavalent vaccine
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "CH_VACC_C_DP1",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::CH_VACC_C_DP1)
}
}
\author{
Qianyu Dong
}
