% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH_VACC_C_MSL.R
\name{CH_VACC_C_MSL}
\alias{CH_VACC_C_MSL}
\title{CH_VACC_C_MSL
KRdata
Measles vaccination received
Percentage of children (age 12-23)}
\usage{
CH_VACC_C_MSL(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
CH_VACC_C_MSL
KRdata
Measles vaccination received
Percentage of children (age 12-23)
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "CH_VACC_C_MSL",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::CH_VACC_C_MSL)
}
}
\author{
Qianyu Dong
}
