% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH_VACS_C_PN3.R
\name{CH_VACS_C_PN3}
\alias{CH_VACS_C_PN3}
\title{CH_VACS_C_PN3
KRdata
Pneumococcal 3 vaccination received}
\usage{
CH_VACS_C_PN3(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
CH_VACS_C_PN3
KRdata
Pneumococcal 3 vaccination received
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "CH_VACS_C_PN3",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::CH_VACS_C_PN3)
}
}
\author{
Qianyu Dong
}
