% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterModel.R
\name{clusterModel}
\alias{clusterModel}
\title{Calculate cluster model estimates using beta binomial model}
\usage{
clusterModel(
  dat.tem,
  cluster.info,
  admin,
  admin.info,
  admin.mat,
  spatialmodel,
  stata
)
}
\arguments{
\item{dat.tem}{dataframe that contains the indicator of interests}

\item{cluster.info}{dataframe that contains admin 1 and admin 2 information and coordinates for each cluster.}

\item{admin}{admin level for the model}

\item{spatialmodel}{model for area-level random effects, "iid" or "bym2.}

\item{stata}{whether or not to include urban/rural stratum.}

\item{admininfo}{dataframe that contains population and urban/rural proportion at specific admin level}
}
\value{
This function returns the dataset that contain district name and population for given  tiff files and polygons of admin level,
  \item { clusterModel.result
}
}
\description{
This function calculate smoothed direct estimates at given admin level.
}
\examples{
\dontrun{
}

}
\author{
Qianyu Dong
}
