% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM_ECMR_C_NNF.R
\name{CM_ECMR_C_NNF}
\alias{CM_ECMR_C_NNF}
\title{CM_ECMR_C_NNF NMR five years prior to survey.}
\usage{
CM_ECMR_C_NNF(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
BR
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "CM_ECMR_C_NNF", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
