% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM_ECMR_C_NNR.R
\name{CM_ECMR_C_NNR}
\alias{CM_ECMR_C_NNR}
\title{CM_ECMR_C_NNR
BRdata
Neonatal mortality rate}
\usage{
CM_ECMR_C_NNR(Rdata, nmr.year = 10)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}

\item{nmr.year}{This is an argument specifically for NMR calculation. It specifies births how many years do we include prior to the date of survey. Default to be 10, i.e., NMR in the last 10 years prior to survey.}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
CM_ECMR_C_NNR
BRdata
Neonatal mortality rate
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "CM_ECMR_C_NNR",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::CM_ECMR_C_NNR)
}
}
\author{
Qianyu Dong
}
