% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM_ECMR_C_U5F.R
\name{CM_ECMR_C_U5F}
\alias{CM_ECMR_C_U5F}
\title{CM_ECMR_C_U5F u5mr 5 years prior to survey.}
\usage{
CM_ECMR_C_U5F(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
BR
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "CM_ECMR_C_U5M", year = 2018)
}

}
