% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM_ECMR_C_U5M.R
\name{CM_ECMR_C_U5M}
\alias{CM_ECMR_C_U5M}
\title{CM_ECMR_C_U5M u5mr 10 years prior to survey.}
\usage{
CM_ECMR_C_U5M(Rdata, mort.cut = c(1, 2, 6, 12, 24, 36, 48, 60), mort.year = 10)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}

\item{mort.cut}{age group cutoffs for hazard calculation. Constant hazards are assumed to be within the specified cutoffs.}

\item{mort.year}{The number of years preceding survey. Set to be 10.}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
BR
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "CM_ECMR_C_U5M", year = 2018)
}

}
