% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CN_ANMC_C_ANY.R
\name{CN_ANMC_C_ANY}
\alias{CN_ANMC_C_ANY}
\title{CN_ANMC_C_ANY
PRdata
Children with any anemia
Children under five with any anemia}
\usage{
CN_ANMC_C_ANY(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
CN_ANMC_C_ANY
PRdata
Children with any anemia
Children under five with any anemia
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "CN_ANMC_C_ANY",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::CN_ANMC_C_ANY)
}
}
\author{
Qianyu Dong
}
