% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CN_BRFI_C_EVR.R
\name{CN_BRFI_C_EVR}
\alias{CN_BRFI_C_EVR}
\title{CN_BRFI_C_EVR}
\usage{
CN_BRFI_C_EVR(KRdata)
}
\arguments{
\item{KRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
CN_BRFI_C_EVR
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "CN_BRFI_C_EVR", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
