% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CN_BRFS_C_EXB.R
\name{CN_BRFS_C_EXB}
\alias{CN_BRFS_C_EXB}
\title{CN_BRFS_C_EXB
KRdata
Children exclusively breastfed
Prevalence of exclusive breastfeeding of children under six months of age}
\usage{
CN_BRFS_C_EXB(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
CN_BRFS_C_EXB
KRdata
Children exclusively breastfed
Prevalence of exclusive breastfeeding of children under six months of age
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "CN_BRFS_C_EXB",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::CN_BRFS_C_EXB)
}
}
\author{
Qianyu Dong
}
