% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CN_NUTS_C_WH3.R
\name{CN_NUTS_C_WH3}
\alias{CN_NUTS_C_WH3}
\title{CN_NUTS_C_WH3}
\usage{
CN_NUTS_C_WH3(PRdata)
}
\arguments{
\item{PRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
CN_NUTS_C_WH3
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "CN_NUTS_C_WH3", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
