% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CO_INUS_W_U12.R
\name{CO_INUS_W_U12}
\alias{CO_INUS_W_U12}
\title{CO_INUS_W_U12}
\usage{
CO_INUS_W_U12(IRdata)
}
\arguments{
\item{IRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
CO_INUS_W_U12
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
