% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CO_MOBB_W_BNK.R
\name{CO_MOBB_W_BNK}
\alias{CO_MOBB_W_BNK}
\title{CO_MOBB_W_BNK}
\usage{
CO_MOBB_W_BNK(IRdata)
}
\arguments{
\item{IRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
CO_MOBB_W_BNK
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
