% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP_BREG_C_CRT.R
\name{CP_BREG_C_CRT}
\alias{CP_BREG_C_CRT}
\title{CP_BREG_C_CRT}
\usage{
CP_BREG_C_CRT(PRdata)
}
\arguments{
\item{PRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
CP_BREG_C_CRT
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
