% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP_BREG_C_REG.R
\name{CP_BREG_C_REG}
\alias{CP_BREG_C_REG}
\title{CP_BREG_C_REG}
\usage{
CP_BREG_C_REG(PRdata)
}
\arguments{
\item{PRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
CP_BREG_C_REG
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
