% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U5MR_helpers.R
\name{directEST_u5mr}
\alias{directEST_u5mr}
\title{Helper function for the U5MR direct estimation}
\usage{
directEST_u5mr(modt, regionVar)
}
\arguments{
\item{modt}{input data frame internally constructed from directEST() function}

\item{regionVar}{region column}
}
\value{
a fitted object in the same format as smoothSurvey() function, to be slotted in directEST() internal script.
}
\description{
Helper function for the U5MR direct estimation
}
