% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV_AFSV_W_A10.R
\name{DV_AFSV_W_A10}
\alias{DV_AFSV_W_A10}
\title{DV_AFSV_W_A10}
\usage{
DV_AFSV_W_A10(IRdata)
}
\arguments{
\item{IRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator

A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
DV_AFSV_W_A10
}
\examples{
\dontrun{
}
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
