% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV_EXPV_W_12M.R
\name{DV_EXPV_W_12M}
\alias{DV_EXPV_W_12M}
\title{DV_EXPV_W_12M  Percentage of women who have experienced physical violence in the past 12 months often or sometimes
dv_phy_12m in github
IR}
\usage{
DV_EXPV_W_12M(Rdata)
}
\arguments{
\item{Rdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
DV_EXPV_W_12M  Percentage of women who have experienced physical violence in the past 12 months often or sometimes
dv_phy_12m in github
IR
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
