% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV_EXSV_W_EVR.R
\name{DV_EXSV_W_EVR}
\alias{DV_EXSV_W_EVR}
\title{DV_EXSV_W_EVR  Percentage of women who ever experienced sexual violence
dv_sex in github
IR}
\usage{
DV_EXSV_W_EVR(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
DV_EXSV_W_EVR  Percentage of women who ever experienced sexual violence
dv_sex in github
IR
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "DV_EXSV_W_EVR", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
