% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV_PCPV_W_EMP.R
\name{DV_PCPV_W_EMP}
\alias{DV_PCPV_W_EMP}
\title{DV_PCPV_W_EMP}
\usage{
DV_PCPV_W_EMP(IRdata)
}
\arguments{
\item{IRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
DV_PCPV_W_EMP
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "DV_PCPV_W_EMP", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
