% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV_PCPV_W_FLW.R
\name{DV_PCPV_W_FLW}
\alias{DV_PCPV_W_FLW}
\title{DV_PCPV_W_FLW}
\usage{
DV_PCPV_W_FLW(IRdata)
}
\arguments{
\item{IRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
DV_PCPV_W_FLW
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
