% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV_STPS_W_BFR.R
\name{DV_STPS_W_BFR}
\alias{DV_STPS_W_BFR}
\title{DV_STPS_W_BFR}
\usage{
DV_STPS_W_BFR(IRdata)
}
\arguments{
\item{IRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
DV_STPS_W_BFR
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "DV_STPS_W_BFR", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
