% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV_VPRG_W_VPG.R
\name{DV_VPRG_W_VPG}
\alias{DV_VPRG_W_VPG}
\title{DV_VPRG_W_VPG}
\usage{
DV_VPRG_W_VPG(IRdata)
}
\arguments{
\item{IRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
DV_VPRG_W_VPG
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
