% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ED_EDUC_W_SEH.R
\name{ED_EDUC_W_SEH}
\alias{ED_EDUC_W_SEH}
\title{ED_EDUC_W_SEH
IRdata
Percentage of women with secondary or higher education}
\usage{
ED_EDUC_W_SEH(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
ED_EDUC_W_SEH
IRdata
Percentage of women with secondary or higher education
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "FP_CUSA_W_MOD",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::fp_cruse_mod)
}
}
\author{
Qianyu Dong
}
