% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ED_LITR_W_LIT.R
\name{ED_LITR_W_LIT}
\alias{ED_LITR_W_LIT}
\title{ED_LITR_W_LIT}
\usage{
ED_LITR_W_LIT(IRdata)
}
\arguments{
\item{IRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
ED_LITR_W_LIT
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
