% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ED_MDIA_W_3MD.R
\name{ED_MDIA_W_3MD}
\alias{ED_MDIA_W_3MD}
\title{ED_MDIA_W_3MD}
\usage{
ED_MDIA_W_3MD(IRdata)
}
\arguments{
\item{IRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
ED_MDIA_W_3MD
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "ED_MDIA_W_3MD", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
