% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ED_MDIA_W_TLV.R
\name{ED_MDIA_W_TLV}
\alias{ED_MDIA_W_TLV}
\title{ED_MDIA_W_TLV}
\usage{
ED_MDIA_W_TLV(IRdata)
}
\arguments{
\item{IRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
ED_MDIA_W_TLV
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
