% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FG_PFCC_W_WCC.R
\name{FG_PFCC_W_WCC}
\alias{FG_PFCC_W_WCC}
\title{FG_PFCC_W_WCC
IRdata
Percentage of women circumcised (women who experienced female genital cutting (FGM))}
\usage{
FG_PFCC_W_WCC(Rdata)
}
\arguments{
\item{Rdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
FG_PFCC_W_WCC
IRdata
Percentage of women circumcised (women who experienced female genital cutting (FGM))
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "FP_CUSA_W_MOD",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::fp_cruse_mod)
}
}
\author{
Qianyu Dong
}
