% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smthdirectEST.R
\name{fhModel}
\alias{fhModel}
\title{Calculate smoothed direct estimates}
\usage{
fhModel(data, cluster.info, admin.info, admin)
}
\arguments{
\item{data}{dataframe that contains the indicator of interests}

\item{cluster.info}{dataframe that contains admin 1 and admin 2 information and coordinates for each cluster.}

\item{admin.info}{dataframe that contains population and urban/rural proportion at specific admin level}

\item{admin}{admin level for the model}
}
\value{
This function returns the dataset that contain district name and population for given  tiff files and polygons of admin level,
  \item { 
}
}
\description{
This function calculate smoothed direct estimates at given admin level.
}
\examples{
\dontrun{
}

}
\author{
Qianyu Dong
}
