% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{fp_cruse_mod}
\alias{fp_cruse_mod}
\title{FP_CUSM_W_MOD
IRdata
Modern contraceptive prevalence rate (Married women currently using any modern method of contraception)}
\usage{
fp_cruse_mod(Rdata)
}
\arguments{
\item{Rdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
FP_CUSM_W_MOD
IRdata
Modern contraceptive prevalence rate (Married women currently using any modern method of contraception)
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "FP_CUSM_W_MOD",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::fp_cruse_mod)
}
}
\author{
Qianyu Dong
}
