% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_CUSA_W_IMP.R
\name{FP_CUSA_W_IMP}
\alias{FP_CUSA_W_IMP}
\title{FP_CUSA_W_IMP}
\usage{
FP_CUSA_W_IMP(IRdata)
}
\arguments{
\item{IRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
FP_CUSA_W_IMP
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
