% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_CUSA_W_INJ.R
\name{FP_CUSA_W_INJ}
\alias{FP_CUSA_W_INJ}
\title{FP_CUSA_W_INJ}
\usage{
FP_CUSA_W_INJ(IRdata)
}
\arguments{
\item{IRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
FP_CUSA_W_INJ
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "FP_CUSA_W_INJ", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
