% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_CUSA_W_STD.R
\name{FP_CUSA_W_STD}
\alias{FP_CUSA_W_STD}
\title{FP_CUSA_W_STD}
\usage{
FP_CUSA_W_STD(IRdata)
}
\arguments{
\item{IRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
FP_CUSA_W_STD
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
