% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_CUSM_W_MOD.R
\name{FP_CUSM_W_MOD}
\alias{FP_CUSM_W_MOD}
\title{FP_CUSM_W_MOD
IRdata
Modern contraceptive prevalence rate (Married women currently using any modern method of contraception)}
\usage{
FP_CUSM_W_MOD(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
FP_CUSM_W_MOD
IRdata
Modern contraceptive prevalence rate (Married women currently using any modern method of contraception)
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "FP_CUSA_W_MOD",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::fp_cruse_mod)
}
}
\author{
Qianyu Dong
}
