% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_EFPM_W_NWS.R
\name{FP_EFPM_W_NWS}
\alias{FP_EFPM_W_NWS}
\title{FP_EFPM_W_NWS}
\usage{
FP_EFPM_W_NWS(IRdata)
}
\arguments{
\item{IRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
FP_EFPM_W_NWS
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
