% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_EVUM_W_MOD.R
\name{FP_EVUM_W_MOD}
\alias{FP_EVUM_W_MOD}
\title{FP_EVUM_W_MOD}
\usage{
FP_EVUM_W_MOD(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
FP_EVUM_W_MOD
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
