% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_KMTA_W_FCN.R
\name{FP_KMTA_W_FCN}
\alias{FP_KMTA_W_FCN}
\title{FP_KMTA_W_FCN}
\usage{
FP_KMTA_W_FCN(IRdata)
}
\arguments{
\item{IRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
FP_KMTA_W_FCN
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
