% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_KMTA_W_OMD.R
\name{FP_KMTA_W_OMD}
\alias{FP_KMTA_W_OMD}
\title{FP_KMTA_W_OMD}
\usage{
FP_KMTA_W_OMD(IRdata)
}
\arguments{
\item{IRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
FP_KMTA_W_OMD
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
