% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_KMTA_W_STD.R
\name{FP_KMTA_W_STD}
\alias{FP_KMTA_W_STD}
\title{FP_KMTA_W_STD}
\usage{
FP_KMTA_W_STD(IRdata)
}
\arguments{
\item{IRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
FP_KMTA_W_STD
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "FP_KMTA_W_STD", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
