% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_NADM_W_UNT.R
\name{FP_NADM_W_UNT}
\alias{FP_NADM_W_UNT}
\title{FP_NADM_W_UNT
#unmet_family
IRdata
Married women with an unmet need for family planning for spacing and limiting, line 17 manually added by Qianyu}
\usage{
FP_NADM_W_UNT(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
FP_NADM_W_UNT
#unmet_family
IRdata
Married women with an unmet need for family planning for spacing and limiting, line 17 manually added by Qianyu
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "FP_NADM_W_UNT", year = 2018)
}

}
\author{
Qianyu Dong
}
