% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_NDYA_W_PDM.R
\name{FP_NDYA_W_PDM}
\alias{FP_NDYA_W_PDM}
\title{FP_NADA_W_UNT
IRdata
Women with an unmet need for family planning for spacing and limiting}
\usage{
FP_NDYA_W_PDM(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
FP_NADA_W_UNT
IRdata
Women with an unmet need for family planning for spacing and limiting
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "FP_NDYA_W_PDM",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::FP_NDYA_W_PDM)
}
}
\author{
Qianyu Dong
}
