% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_api_table.R
\name{get_api_table}
\alias{get_api_table}
\title{Function to obtain subnational estimates from DHS API}
\usage{
get_api_table(coutry, survey, indicator, simplify = TRUE)
}
\arguments{
\item{survey}{keys at: https://api.dhsprogram.com/rest/dhs/surveys?returnFields=SurveyId,SurveyYearLabel,SurveyType,CountryName&f=html}

\item{indicator}{keys at: https://api.dhsprogram.com/rest/dhs/indicators?returnFields=IndicatorId,Label,Definition&f=html}

\item{simplify}{if TRUE only the value and region index is returned.}

\item{country}{keys at: https://api.dhsprogram.com/rest/dhs/countries?returnFields=CountryName,DHS_CountryCode&f=html}
}
\description{
Function to obtain subnational estimates from DHS API
}
