% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HA_AFSY_M_A15.R
\name{HA_AFSY_M_A15}
\alias{HA_AFSY_M_A15}
\title{HA_AFSY_M_A15}
\usage{
HA_AFSY_M_A15(MRdata)
}
\arguments{
\item{MRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
HA_AFSY_M_A15
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "HA_AFSY_M_A15", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
