% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HA_ANSS_W_RSX.R
\name{HA_ANSS_W_RSX}
\alias{HA_ANSS_W_RSX}
\title{HA_ANSS_W_RSX}
\usage{
HA_ANSS_W_RSX(IRdata)
}
\arguments{
\item{IRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
HA_ANSS_W_RSX
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
