% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HA_CATH_W_ATN.R
\name{HA_CATH_W_ATN}
\alias{HA_CATH_W_ATN}
\title{HA_CATH_W_ATN}
\usage{
HA_CATH_W_ATN(IRdata)
}
\arguments{
\item{IRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator

A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator

A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
HA_CATH_W_ATN
}
\examples{
\dontrun{
}
\dontrun{
}
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
