% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HA_HIVP_B_HIV.R
\name{HA_HIVP_B_HIV}
\alias{HA_HIVP_B_HIV}
\title{HA_HIVP_B_HIV
hv_hiv_pos
"HIV prevalence among general population"}
\usage{
HA_HIVP_B_HIV(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata IR+AR+MR}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
HA_HIVP_B_HIV
hv_hiv_pos
"HIV prevalence among general population"
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "HA_HIVP_B_HIV",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::HA_HIVP_B_HIV)
}
}
\author{
Qianyu Dong
}
