% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HA_HRSX_W_CND.R
\name{HA_HRSX_W_CND}
\alias{HA_HRSX_W_CND}
\title{HA_HRSX_W_CND}
\usage{
HA_HRSX_W_CND(IRdata)
}
\arguments{
\item{IRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
HA_HRSX_W_CND
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
