% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HA_HVTY_M_TRR.R
\name{HA_HVTY_M_TRR}
\alias{HA_HVTY_M_TRR}
\title{HA_HVTY_M_TRR}
\usage{
HA_HVTY_M_TRR(MRdata)
}
\arguments{
\item{MRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
HA_HVTY_M_TRR
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
