% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HA_STIS_M_DIS.R
\name{HA_STIS_M_DIS}
\alias{HA_STIS_M_DIS}
\title{HA_STIS_M_DIS}
\usage{
HA_STIS_M_DIS(MRdata)
}
\arguments{
\item{MRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
HA_STIS_M_DIS
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
