% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HA_STIS_M_SOR.R
\name{HA_STIS_M_SOR}
\alias{HA_STIS_M_SOR}
\title{HA_STIS_M_SOR}
\usage{
HA_STIS_M_SOR(MRdata)
}
\arguments{
\item{MRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
HA_STIS_M_SOR
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "HA_STIS_M_SOR", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
