% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HC_AGEG_P_ADL.R
\name{HC_AGEG_P_ADL}
\alias{HC_AGEG_P_ADL}
\title{HC_AGEG_P_ADL}
\usage{
HC_AGEG_P_ADL(PRdata)
}
\arguments{
\item{PRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
HC_AGEG_P_ADL
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "HC_AGEG_P_ADL", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
