% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HC_ELEC_H_ELC.R
\name{HC_ELEC_H_ELC}
\alias{HC_ELEC_H_ELC}
\title{HC_ELEC_H_ELC Percentage of households with electricity
ph_electric in github
HR}
\usage{
HC_ELEC_H_ELC(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
HC_ELEC_H_ELC Percentage of households with electricity
ph_electric in github
HR
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
