% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HC_HEFF_H_RDO.R
\name{HC_HEFF_H_RDO}
\alias{HC_HEFF_H_RDO}
\title{HC_HEFF_H_RDO Percentage of households possessing a radio
ph_radio in github
HR}
\usage{
HC_HEFF_H_RDO(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
HC_HEFF_H_RDO Percentage of households possessing a radio
ph_radio in github
HR
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "HC_HEFF_H_RDO", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
