% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HC_WIXQ_P_12Q.R
\name{HC_WIXQ_P_12Q}
\alias{HC_WIXQ_P_12Q}
\title{HC_WIXQ_P_12Q
PRdata
Population in the lowest and second wealth quintile}
\usage{
HC_WIXQ_P_12Q(Rdata)
}
\arguments{
\item{Rdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
HC_WIXQ_P_12Q
PRdata
Population in the lowest and second wealth quintile
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "FP_CUSA_W_MOD",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::fp_cruse_mod)
}
}
\author{
Qianyu Dong
}
