% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HC_WIXQ_P_2ND.R
\name{HC_WIXQ_P_2ND}
\alias{HC_WIXQ_P_2ND}
\title{HC_WIXQ_P_2ND
PRdata
Population in the second wealth quintile}
\usage{
HC_WIXQ_P_2ND(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
HC_WIXQ_P_2ND
PRdata
Population in the second wealth quintile
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "FP_CUSA_W_MOD",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::fp_cruse_mod)
}
}
\author{
Qianyu Dong
}
