% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{hv_hiv_pos}
\alias{hv_hiv_pos}
\title{HA_HIVP_W_HIV
hv_hiv_pos
"HIV positive test result"}
\usage{
hv_hiv_pos(Rdata)
}
\arguments{
\item{Rdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
HA_HIVP_W_HIV
hv_hiv_pos
"HIV positive test result"
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "HA_HIVP_W_HIV",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::hv_hiv_pos)
}
}
\author{
Qianyu Dong
}
