% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MA_MBAG_W_B15.R
\name{MA_MBAG_W_B15}
\alias{MA_MBAG_W_B15}
\title{MA_MBAG_W_B15  Women first married by exact age 15
ms_afm_15 in github
IR}
\usage{
MA_MBAG_W_B15(Rdata)
}
\arguments{
\item{Rdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
MA_MBAG_W_B15  Women first married by exact age 15
ms_afm_15 in github
IR
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
