% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MA_MBAY_W_B18.R
\name{MA_MBAY_W_B18}
\alias{MA_MBAY_W_B18}
\title{MA_MBAY_W_B18  Percentage of Young women age 20-24 first married by exact age 18
ms_afm_18 in github
IR}
\usage{
MA_MBAY_W_B18(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
MA_MBAY_W_B18  Percentage of Young women age 20-24 first married by exact age 18
ms_afm_18 in github
IR
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
