% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MA_MSTA_W_NMA.R
\name{MA_MSTA_W_NMA}
\alias{MA_MSTA_W_NMA}
\title{MA_MSTA_W_NMA}
\usage{
MA_MSTA_W_NMA(IRdata)
}
\arguments{
\item{IRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
MA_MSTA_W_NMA
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "MA_MSTA_W_NMA", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
