% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_all_result.R
\docType{data}
\name{match_all_result}
\alias{match_all_result}
\title{Indicator matching table.}
\format{
An object of class \code{data.frame} with 146 rows and 21 columns.
}
\usage{
data(match_all_result)
}
\description{
A data frame with DHS indicator information used to match across different code bases.
}
\keyword{datasets}
