% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_NETC_C_ITN.R
\name{ML_NETC_C_ITN}
\alias{ML_NETC_C_ITN}
\title{ML_NETC_C_ITN
PRdata
Percentage of children under age five who slept under an insecticide treated net (ITN) the night before the survey}
\usage{
ML_NETC_C_ITN(Rdata)
}
\arguments{
\item{Rdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
ML_NETC_C_ITN
PRdata
Percentage of children under age five who slept under an insecticide treated net (ITN) the night before the survey
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "FP_CUSA_W_MOD",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::fp_cruse_mod)
}
}
\author{
Qianyu Dong
}
