% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_NETP_H_IT2.R
\name{ML_NETP_H_IT2}
\alias{ML_NETP_H_IT2}
\title{ML_NETP_H_IT2
HRdata
Households with access to an insecticide-treated mosquito net (ITN)}
\usage{
ML_NETP_H_IT2(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
ML_NETP_H_IT2
HRdata
Households with access to an insecticide-treated mosquito net (ITN)
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "ML_NETP_H_IT2",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::ML_NETP_H_IT2)
}
}
\author{
Qianyu Dong
}
