% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_NETU_P_ITN.R
\name{ML_NETU_P_ITN}
\alias{ML_NETU_P_ITN}
\title{ML_NETU_P_ITN
PRdata
Percentage of the de facto household population who slept under an insecticide treated net the night before the survey}
\usage{
ML_NETU_P_ITN(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
ML_NETU_P_ITN
PRdata
Percentage of the de facto household population who slept under an insecticide treated net the night before the survey
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "FP_CUSA_W_MOD",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::fp_cruse_mod)
}
}
\author{
Qianyu Dong
}
