% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_PMAL_C_RDT.R
\name{ML_PMAL_C_RDT}
\alias{ML_PMAL_C_RDT}
\title{ML_PMAL_C_RDT
PRdata
Malaria prevalence according to RDT}
\usage{
ML_PMAL_C_RDT(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
ML_PMAL_C_RDT
PRdata
Malaria prevalence according to RDT
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "ML_PMAL_C_RDT",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::ML_PMAL_C_RDT)
}
}
\author{
Qianyu Dong
}
