% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{NMR}
\alias{NMR}
\title{CM_ECMR_C_NNR
nmr CM_ECMR_C_NNR BR (not from dhs github)
Neonatal mortality rate !!!!!!}
\usage{
NMR(Rdata, nmr.year)
}
\arguments{
\item{Rdata}{data.frame from survryPrev::getDHSdata}

\item{nmr.year}{This is an argument specifically for NMR calculation. It specifies births how many years do we include prior to the date of survey. Default to be 10, i.e., NMR in the last 10 years prior to survey.}
}
\value{
A partially processed data.frame that will be used in  survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
CM_ECMR_C_NNR
nmr CM_ECMR_C_NNR BR (not from dhs github)
Neonatal mortality rate !!!!!!
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "CM_ECMR_C_NNR",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::NMR)
}
}
\author{
Qianyu Dong
}
