% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{nt_ch_any_anem}
\alias{nt_ch_any_anem}
\title{CN_ANMC_C_ANY
Children with any anemia
"Any anemia - child 6-59 months" PR NT_CH_NUT.do
Children under five with any anemia}
\usage{
nt_ch_any_anem(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
CN_ANMC_C_ANY
Children with any anemia
"Any anemia - child 6-59 months" PR NT_CH_NUT.do
Children under five with any anemia
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "CN_ANMC_C_ANY",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::nt_ch_any_anem)
}
}
\author{
Qianyu Dong
}
