% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{nt_wm_any_anem}
\alias{nt_wm_any_anem}
\title{AN_ANEM_W_ANY
womananemia
nt_wm_any_anem "Any anemia - women"  NT_WM_NUT.do
Percentage of women aged 15-49 classified as having any anemia}
\usage{
nt_wm_any_anem(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
AN_ANEM_W_ANY
womananemia
nt_wm_any_anem "Any anemia - women"  NT_WM_NUT.do
Percentage of women aged 15-49 classified as having any anemia
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "AN_ANEM_W_ANY",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::nt_wm_any_anem)
}
}
\author{
Qianyu Dong
}
