% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{nt_wm_thin}
\alias{nt_wm_thin}
\title{AN_NUTS_W_THN
Women who are thin according to BMI (<18.5)
NT_WM_NUT.do "Thin BMI - women" IR !!!!!!!!
Underweight (Prevalence of underweight (BMI < 18.5) women of reproductive age)}
\usage{
nt_wm_thin(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
AN_NUTS_W_THN
Women who are thin according to BMI (<18.5)
NT_WM_NUT.do "Thin BMI - women" IR !!!!!!!!
Underweight (Prevalence of underweight (BMI < 18.5) women of reproductive age)
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "AN_NUTS_W_THN",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::nt_wm_thin)
}
}
\author{
Qianyu Dong
}
