% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{ph_sani_basic}
\alias{ph_sani_basic}
\title{WS_TLET_P_BAS
Population with access to a basic sanitation service WS_TLET_P_BAS in DHS API
PH_SANI.do  PR
ph_sani_basic "Basic sanitation facility"}
\usage{
ph_sani_basic(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
WS_TLET_P_BAS
Population with access to a basic sanitation service WS_TLET_P_BAS in DHS API
PH_SANI.do  PR
ph_sani_basic "Basic sanitation facility"
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "WS_TLET_P_BAS",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::ph_sani_basic)
}
}
\author{
Qianyu Dong
}
