% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visualization.R
\name{prevMap}
\alias{prevMap}
\title{static prevalence map for any subnational level}
\usage{
prevMap(
  res.obj,
  poly.shp,
  admin1.focus = NULL,
  value.to.plot = "mean",
  map.title = NULL,
  legend.label = "Estimates",
  threshold.p = NULL,
  color.palette = NULL,
  color.reverse = T,
  value.range = NULL,
  ...
)
}
\arguments{
\item{res.obj}{result object from surveyPrev}

\item{poly.shp}{polygon file for plotting}

\item{admin1.focus}{if needed, the admin 1 level area name to create an individual focused prevalence map on}

\item{value.to.plot}{statistics to appear on the map}

\item{map.title}{title for the map}

\item{legend.label}{label for the legend, such as 'Coefficient of Variation'}

\item{threshold.p}{cutoff for the exceedance probability map}

\item{color.palette}{which palette to use for plotting}

\item{color.reverse}{whether to use reverse color scale}

\item{value.range}{what range to plot, useful if want to compare plots using the same scale}
}
\value{
summer map object
}
\description{
produce static map for any subnational level, with the option to focus on one admin-1 level
}
\examples{
\dontrun{
geo <- getDHSgeo(country = "Zambia", year = 2018)
cluster.info <- clusterInfo(geo = geo,
                            poly.adm1 = ZambiaAdm1,
                            poly.adm2 = ZambiaAdm2)
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "ancvisit4+",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = "ancvisit4+")
admin.info2 <- adminInfo(poly.adm = sf::st_as_sf(ZambiaAdm2), 
                         admin = 2, 
                         by.adm = "NAME_2", 
                         by.adm.upper = "NAME_1")
cl_res_ad2 <- clusterModel(data=data,
                  cluster.info = cluster.info,
                  admin.info = admin.info2,
                  model = "bym2",
                  admin = 2)

prevMap.adm2.central <- prevMap(res.obj=cl_res_ad2, 
                                       poly.shp=poly.adm2,
                                       admin1.focus = "Central")
} 

}
