% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visualization.R
\name{prevMap.web}
\alias{prevMap.web}
\title{leaflet (interactive) prevalence map for any subnational level}
\usage{
prevMap.web(
  res.obj,
  poly.shp,
  admin1.focus = NULL,
  color.palette = NULL,
  value.to.plot = "mean",
  value.range = NULL,
  num_bins = NULL,
  legend.label = "Estimates",
  map.title = NULL,
  color.reverse = T,
  no.hatching = F,
  hatching.density = 12,
  use.basemap = NULL,
  threshold.p = NULL,
  legend.color.reverse = T
)
}
\arguments{
\item{res.obj}{result object from surveyPrev}

\item{poly.shp}{polygon file for plotting}

\item{admin1.focus}{if needed, the admin 1 level area name to create an individual focused prevalence map on}

\item{color.palette}{which palette to use for plotting}

\item{value.to.plot}{statistics to appear on the map}

\item{value.range}{what range to plot, useful if want to compare plots using the same scale}

\item{num_bins}{number of bins on the legend (what displayed might not be exact)}

\item{legend.label}{label for the legend, such as 'Coefficient of Variation'}

\item{map.title}{title for the map}

\item{color.reverse}{whether to use reverse color scale}

\item{no.hatching}{whether to hatch region with problematic uncertainties, recommend F (depends on rgeos package, set to T if not installed)}

\item{hatching.density}{density of the hatching lines}

\item{use.basemap}{what basemap to use 'OSM', if NULL, no basemap}

\item{threshold.p}{cutoff for the exceedance probability map}

\item{legend.color.reverse}{logical indicator of reversing the color legend}
}
\value{
leaflet map object
}
\description{
produce interactive map for any subnational level, with the option to focus on one admin-1 level
}
\examples{
\dontrun{
geo <- getDHSgeo(country = "Zambia", year = 2018)
cluster.info <- clusterInfo(geo = geo,
                            poly.adm1 = ZambiaAdm1,
                            poly.adm2 = ZambiaAdm2)
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "ancvisit4+",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = "ancvisit4+")
poly.adm2 <- sf::st_as_sf(ZambiaAdm2)
admin.info2 <- adminInfo(poly.adm = poly.adm2, 
                         admin = 2, 
                         by.adm = "NAME_2", 
                         by.adm.upper = "NAME_1")
cl_res_ad2 <- clusterModel(data=data,
                  cluster.info = cluster.info,
                  admin.info = admin.info2,
                  model = "bym2",
                  admin = 2)
prevMap.adm2.central <- prevMap.web(res.obj=cl_res_ad2, 
                                              poly.shp=poly.adm2,
                                              admin1.focus = "Central")
} 

}
