% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankPlot.R
\name{rankPlot}
\alias{rankPlot}
\title{Get ranking plot of model results}
\usage{
rankPlot(x, direction = 1)
}
\arguments{
\item{x}{a model result  using surveyPrev of class "fhModel" or "clusterModel"}

\item{direction}{direction of the ranking. The default is direction = 1, which correspond to larger value having higher ranking. If direction = -1, larger value has lower ranking.}
}
\value{
This function returns a ranking plot.
}
\description{
This function return scatter plot at admin 1 or 2 level for given model results.
}
\examples{
\dontrun{

geo <- getDHSgeo(country = "Zambia", year = 2018)
data(ZambiaAdm1)
data(ZambiaAdm2)
data(ZambiaPopWomen)
cluster.info <- clusterInfo(geo = geo,
                            poly.adm1 = ZambiaAdm1,
                            poly.adm2 = ZambiaAdm2)

dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "ancvisit4+",
                                 year = 2018)

data <- getDHSindicator(dhsData, indicator = "ancvisit4+")
admin.info2 <- adminInfo(poly.adm = ZambiaAdm2,
                        admin = 2,
						   by.adm="NAME_2",
						   by.adm.upper = "NAME_1")
cl_res_ad2_unstrat <- clusterModel(data = data,
                  cluster.info = cluster.info,
                  admin.info = admin.info2,
                  stratification = FALSE,
                  model = "bym2",
                  admin = 2,
                  aggregation = TRUE,
                  CI = 0.95)
rankPlot(cl_res_ad2_unstrat)

}

}
\author{
Zehang Li, Qianyu Dong
}
