% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{rh_anc_4vs}
\alias{rh_anc_4vs}
\title{RH_ANCN_W_N4P
ancvisit4+ RH_ANCN_W_N4P  IR
Antenatal visits for pregnancy: 4+ visits}
\usage{
rh_anc_4vs(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
RH_ANCN_W_N4P
ancvisit4+ RH_ANCN_W_N4P  IR
Antenatal visits for pregnancy: 4+ visits
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "RH_ANCN_W_N4P",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::rh_anc_4vs)
}
}
\author{
Qianyu Dong
}
