% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RH_ANCN_W_N01.R
\name{RH_ANCN_W_N01}
\alias{RH_ANCN_W_N01}
\title{AH_TOBC_W_OTH}
\usage{
RH_ANCN_W_N01(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
AH_TOBC_W_OTH
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
