% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RH_ANCN_W_NON.R
\name{RH_ANCN_W_NON}
\alias{RH_ANCN_W_NON}
\title{RH_ANCN_W_N4P
IRdata
Antenatal visits for pregnancy: no visits TWO year}
\usage{
RH_ANCN_W_NON(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
RH_ANCN_W_N4P
IRdata
Antenatal visits for pregnancy: no visits TWO year
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "RH_ANCN_W_NON",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::RH_ANCN_W_NON)
}
}
\author{
Qianyu Dong
}
