% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{rh_del_pvskill}
\alias{rh_del_pvskill}
\title{RH_DELA_C_SKP
IR or BR
Assistance during delivery from a skilled provider}
\usage{
rh_del_pvskill(Rdata)
}
\arguments{
\item{Rdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
RH_DELA_C_SKP
IR or BR
Assistance during delivery from a skilled provider
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "RH_DELA_C_SKP",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::rh_del_pvskill)
}
}
\author{
Qianyu Dong
}
