% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RH_DELA_C_SKF.R
\name{RH_DELA_C_SKF}
\alias{RH_DELA_C_SKF}
\title{RH_DELA_C_SKF
BRdata
Assistance during delivery from a skilled provider}
\usage{
RH_DELA_C_SKF(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
RH_DELA_C_SKF
BRdata
Assistance during delivery from a skilled provider
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "RH_DELA_C_SKF",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::RH_DELA_C_SKF)
}
}
\author{
Qianyu Dong
}
