% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RH_DELP_C_PUB.R
\name{RH_DELP_C_PUB}
\alias{RH_DELP_C_PUB}
\title{RH_DELP_C_PUB}
\usage{
RH_DELP_C_PUB(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
RH_DELP_C_PUB
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "RH_DELP_C_PUB", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
