% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RH_PAHC_W_MON.R
\name{RH_PAHC_W_MON}
\alias{RH_PAHC_W_MON}
\title{RH_PAHC_W_MON
IRdata
Problems in accessing health care: Getting money for treatment}
\usage{
RH_PAHC_W_MON(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
RH_PAHC_W_MON
IRdata
Problems in accessing health care: Getting money for treatment
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "FP_CUSA_W_MOD",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::fp_cruse_mod)
}
}
\author{
Qianyu Dong
}
