% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RH_PCCT_C_DY2.R
\name{RH_PCCT_C_DY2}
\alias{RH_PCCT_C_DY2}
\title{RH_PCCT_C_DY2 Newborn's first postnatal checkup: 1-2 days
rh_pnc_nb_2days in github
IR}
\usage{
RH_PCCT_C_DY2(Rdata)
}
\arguments{
\item{Rdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
RH_PCCT_C_DY2 Newborn's first postnatal checkup: 1-2 days
rh_pnc_nb_2days in github
IR
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
