% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RH_PCMT_W_NON.R
\name{RH_PCMT_W_NON}
\alias{RH_PCMT_W_NON}
\title{RH_PCMT_W_NON  No Mother's postnatal checkup in 42 days
IR}
\usage{
RH_PCMT_W_NON(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
RH_PCMT_W_NON  No Mother's postnatal checkup in 42 days
IR
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "RH_PCMT_W_NON", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
