% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visualization.R
\name{ridgeprevPlot}
\alias{ridgeprevPlot}
\title{ridge plot for prevalence}
\usage{
ridgeprevPlot(
  res.obj,
  admin1.focus = NA,
  plot.extreme.num = 8,
  legend.label = "Value",
  color.reverse = T,
  plot.format = c("Long", "Wide")[1],
  top.bottom.label = c("Top", "Bottom")
)
}
\arguments{
\item{res.obj}{result object from surveyPrev}

\item{admin1.focus}{whether to plot densities for regions within a single upper admin}

\item{plot.extreme.num}{number of regions to plot for the top n and bottom n regions}

\item{legend.label}{label on the legend}

\item{color.reverse}{whether to reverse color scheme}

\item{plot.format}{c('Long','Wide') for extreme regions, side-by-side or long plot}

\item{top.bottom.label}{c('Top','Bottom') how to name the extremes, top 10 bottom 10? need to change when close to 0 is bad for the indicator}
}
\value{
ggplot2 object
}
\description{
static ridge plot for posterior densities
}
\examples{
\dontrun{
geo <- getDHSgeo(country = "Zambia", year = 2018)
cluster.info <- clusterInfo(geo = geo,
                            poly.adm1 = ZambiaAdm1,
                            poly.adm2 = ZambiaAdm2)
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "ancvisit4+",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = "ancvisit4+")
poly.adm1 <- sf::st_as_sf(ZambiaAdm1)
admin.info1 <- adminInfo(poly.adm = poly.adm1, 
                         admin = 1, 
                         by.adm = "NAME_1")
cl_res_ad1 <- clusterModel(data=data,
                  cluster.info = cluster.info,
                  admin.info = admin.info1,
                  stratification = FALSE,
                  model = "bym2",
                  admin = 1,
                  aggregation = TRUE,
                  CI = 0.95)
 
ridgeprevPlot(cl_res_ad1, plot.extreme.num=5)

}

}
