% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlot.R
\name{scatterPlot}
\alias{scatterPlot}
\title{Get scatter plot for any two model results}
\usage{
scatterPlot(
  res1,
  value1,
  res2,
  value2,
  label1,
  label2,
  by.res1,
  by.res2,
  title
)
}
\arguments{
\item{res1}{model result 1 using surveyPrev}

\item{value1}{value1}

\item{res2}{model result 2 using surveyPrev}

\item{value2}{value2}

\item{label1}{label for x axis}

\item{label2}{label for y axis}

\item{by.res1}{by.res1}

\item{by.res2}{by.res2}

\item{title}{title}
}
\value{
This function returns the dataset that contain district name and population for given  tiff files and polygons of admin level
  \item {plot
}
}
\description{
This function return scatter plot at admin 1 level for any two model results
}
\examples{
\dontrun{

}

}
\author{
Qianyu Dong
}
