% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smthdirectEST.R
\name{smthdirectEST}
\alias{smthdirectEST}
\title{Calculate smoothed direct estimates}
\usage{
smthdirectEST(dat.tem, cluserinfo, admininfo, admin, Amat)
}
\arguments{
\item{dat.tem}{dataframe that contains the indicator of interests}

\item{cluserinfo}{dataframe that contains admin 1 and admin 2 information and coordinates for each cluster.}

\item{admininfo}{dataframe that contains population and urban/rural proportion at specific admin level}

\item{admin}{admin level for the model}
}
\value{
This function returns the dataset that contain district name and population for given  tiff files and polygons of admin level,
  \item { smthdirectEST.result
}
}
\description{
This function calculate smoothed direct estimates at given admin level.
}
\examples{
\dontrun{
}

}
\author{
Qianyu Dong
}
