% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyPrevIndicators.R
\docType{data}
\name{surveyPrevIndicators}
\alias{surveyPrevIndicators}
\title{Table of built-in indicators.}
\format{
An object of class \code{data.frame} with 22 rows and 3 columns.
}
\usage{
data(surveyPrevIndicators)
}
\description{
A data frame of indicators currently implemented in the package
}
\keyword{datasets}
