% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{watersource_adj}
\alias{watersource_adj}
\title{WS_SRCE_P_BAS
Population using a basic water source
PH_WATER.do
ph_wtr_basic "Basic water service" PR}
\usage{
watersource_adj(Rdata)
}
\arguments{
\item{Rdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
WS_SRCE_P_BAS
Population using a basic water source
PH_WATER.do
ph_wtr_basic "Basic water service" PR
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "WS_SRCE_P_BAS",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::watersource_adj)
}
}
\author{
Qianyu Dong
}
