% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WE_AWBT_M_BFD.R
\name{WE_AWBT_M_BFD}
\alias{WE_AWBT_M_BFD}
\title{WE_AWBT_M_BFD}
\usage{
WE_AWBT_M_BFD(MRdata)
}
\arguments{
\item{MRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
WE_AWBT_M_BFD
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
