% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WE_AWBT_M_REF.R
\name{WE_AWBT_M_REF}
\alias{WE_AWBT_M_REF}
\title{WE_AWBT_M_REF}
\usage{
WE_AWBT_M_REF(MRdata)
}
\arguments{
\item{MRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
WE_AWBT_M_REF
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "WE_AWBT_M_REF", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
