% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WE_AWBT_W_BFD.R
\name{WE_AWBT_W_BFD}
\alias{WE_AWBT_W_BFD}
\title{WE_AWBT_W_BFD}
\usage{
WE_AWBT_W_BFD(IRdata)
}
\arguments{
\item{IRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
WE_AWBT_W_BFD
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "WE_AWBT_W_BFD", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
