% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WE_AWBT_W_NEG.R
\name{WE_AWBT_W_NEG}
\alias{WE_AWBT_W_NEG}
\title{WE_AWBT_W_NEG}
\usage{
WE_AWBT_W_NEG(IRdata)
}
\arguments{
\item{IRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
WE_AWBT_W_NEG
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
