% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WE_AWBT_W_REF.R
\name{WE_AWBT_W_REF}
\alias{WE_AWBT_W_REF}
\title{WE_AWBT_W_REF}
\usage{
WE_AWBT_W_REF(IRdata)
}
\arguments{
\item{IRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
WE_AWBT_W_REF
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
