% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WS_SRCE_P_BAS.R
\name{WS_SRCE_P_BAS}
\alias{WS_SRCE_P_BAS}
\title{WS_SRCE_P_BAS
PRdata
Population using a basic water source}
\usage{
WS_SRCE_P_BAS(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
WS_SRCE_P_BAS
PRdata
Population using a basic water source
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "WS_SRCE_P_BAS",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::WS_SRCE_P_BAS)
}
}
\author{
Qianyu Dong
}
