% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WS_TLET_H_IMP.R
\name{WS_TLET_H_IMP}
\alias{WS_TLET_H_IMP}
\title{WS_TLET_H_IMP
PRdata
Percentage of households using an improved sanitation facility}
\usage{
WS_TLET_H_IMP(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
WS_TLET_H_IMP
PRdata
Percentage of households using an improved sanitation facility
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "WS_TLET_H_IMP",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::WS_TLET_H_IMP)
}
}
\author{
Qianyu Dong
}
