% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZambiaAdm1.R
\docType{data}
\name{ZambiaAdm1}
\alias{ZambiaAdm1}
\title{Admin 1 Polygon Map for Zambia.}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 10 rows and 11 columns.
}
\usage{
data(ZambiaAdm1)
}
\description{
A SpatialPolygonsDataFrame corresponding to Zambia's admin-1 regions. The dataset is downloaded from GADM (https://gadm.org/data.html) version 4.1.
}
\keyword{datasets}
