% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZambiaAdm2.R
\docType{data}
\name{ZambiaAdm2}
\alias{ZambiaAdm2}
\title{Admin 2 Polygon Map for Zambia.}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 115 rows and 13 columns.
}
\usage{
data(ZambiaAdm2)
}
\description{
A SpatialPolygonsDataFrame corresponding to Zambia's admin-2 regions. The dataset is downloaded from GADM (https://gadm.org/data.html) version 4.1.
}
\keyword{datasets}
