% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZambiaPopWomen.R
\docType{data}
\name{ZambiaPopWomen}
\alias{ZambiaPopWomen}
\title{Population estimates for Women of age 15 to 49 in Zambia in 2018.}
\format{
An object of class \code{list} of length 4.
}
\usage{
data(ZambiaPopWomen)
}
\description{
A list of three objects
\itemize{
\item raster A 100m by 100m raster file for the population estimates for women of age 15 to 49 in Zambia in 2018
\item admin1_urban A data frame specifying the proportion of urban population (as defined by those living in regions designated as urban in the previous census) for Women of age 15 to 49 in Zambia in 2018, in each admin1 region.
\item admin2_urban A data frame specifying the proportion of urban population (as defined by those living in regions designated as urban in the previous census) for Women of age 15 to 49 in Zambia in 2018, in each admin2 region. The corresponding admin1 region name is also included. 
}
The dataset is downloaded from WorldPop (https://hub.worldpop.org/geodata/summary?id=16429) and post processed.
}
\keyword{datasets}
