% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNewIndicator.R
\name{addNewIndicator}
\alias{addNewIndicator}
\title{Create new row to add in indicatorList.}
\usage{
addNewIndicator(
  new_id,
  dictionary,
  IR = FALSE,
  PR = FALSE,
  KR = FALSE,
  BR = FALSE,
  HR = FALSE,
  MR = FALSE,
  AR = FALSE,
  CR = FALSE
)
}
\arguments{
\item{new_id}{The ID of the new indicator (e.g., "FE_FRTR_W_A10").}

\item{dictionary}{The DHS indicator dictionary data frame, e.g.sae4health::DHS_ind_dictionary}

\item{chapters}{The DHS chapters reference data frame.}

\item{...}{Boolean flags for IR, PR, KR, BR, HR, MR, AR, CR.}
}
\value{
The a row to rbind into indicator_df
}
\description{
Create new row to add in indicatorList.
}
\examples{
\dontrun{
newrow=addNewIndicator(new_id="RH_DELP_C_HOM", dictionary= sae4health::DHS_ind_dictionary,
IR=FALSE, PR=FALSE, KR=FALSE, BR=TRUE,
HR=FALSE, MR=FALSE, AR=FALSE, CR=FALSE )
newrow <- newrow[, names(indicatorList)]
indicatorList <- rbind(indicatorList, newrow)
}


}
\author{
Qianyu Dong
}
