% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admininfo.R
\name{adminInfo}
\alias{adminInfo}
\title{Get admin information}
\usage{
adminInfo(agg.pop, SpatialPolygons, proportion, admin)
}
\arguments{
\item{agg.pop}{data frame of aggregated poplulation from aggPopulation function. It should have two columns: "DistrictName" and "population".}

\item{SpatialPolygons}{spatial polygons dataframe for admin 1 or admin 2.}

\item{proportion}{data frame of urban/rural proportions. For admin1, is should have two columns: "admin1.name" and "urban". For admin2, it should have three columns: "admin1.name", "admin2.name", and "urban", in order to avoid issues merging datasets with duplicated admin2 names.}

\item{admin}{admin level}
}
\value{
This function returns the 1. dataframe that contains admin 1 and admin 2 information and coordinates for each cluster and 2. Adjacency matrix
\itemize{
  \item admin.info
}
}
\description{
This function get admin information including name, character, population and unban/rural proportion.
}
\examples{
\dontrun{
}

}
\author{
Qianyu Dong
}
