% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admininfo.R
\name{adminInfo}
\alias{adminInfo}
\title{Get admin information}
\usage{
adminInfo(geo, admin, agg.pop = NULL, agg.survey = NULL, proportion = NULL)
}
\arguments{
\item{geo}{spatial polygons dataframe for admin 1 or admin 2. This object can be either an sp::SpatialPolygonsDataFrame object or an sf object.}

\item{admin}{admin level}

\item{agg.survey}{data frame of aggregated poplulation from aggPopulation function. It should have two columns: "admin2.name.full" and "population".}

\item{proportion}{data frame of urban/rural proportions. For admin1, is should have two columns: "admin1.name" and "urban". For admin2, it should have three columns: "admin1.name", "admin2.name", and "urban", in order to avoid issues merging datasets with duplicated admin2 names.}
}
\value{
This function returns the 1. dataframe that contains admin 1 and admin 2 information and coordinates for each cluster and 2. Adjacency matrix.
}
\description{
This function get admin information including name, character, population and unban/rural proportion.
}
\examples{

# For sp::SpatialPolygonsDataFrame object
data(ZambiaAdm1)
class(ZambiaAdm1)
info <- adminInfo(ZambiaAdm1, admin = 1)

# For sf object
geo.sf <- sf::st_as_sf(ZambiaAdm1)
info <- adminInfo(geo.sf, admin = 1)

# To include the population information
data(ZambiaPopWomen)
info <- adminInfo(geo = ZambiaAdm1,
                  admin = 1,
                  agg.pop = ZambiaPopWomen$admin1_pop,
                  proportion = ZambiaPopWomen$admin1_urban,
                  agg.survey = agg.survey1)
}
\author{
Qianyu Dong
}
