% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggSurveyWeight.R
\name{aggSurveyWeight}
\alias{aggSurveyWeight}
\title{Get survey weight by admin levels}
\usage{
aggSurveyWeight(data, cluster.info, admin, admin2.name.full = NULL)
}
\arguments{
\item{data}{factor to aggregate pixels. Default to be 10, i.e., the population estimates will be saved on 1km by 1km grids if the input is 100m by 100m tiff. Larger values of aggregation factor improves the computation speed, but can introduce more errors when the regions defined by the polygon are small in size.}

\item{cluster.info}{outcome from clusterinfo}

\item{admin.info}{outcome from admininfo}
}
\value{
This function returns the dataset that contain admin name and survey weight.
}
\description{
This function aggregate survey weight to particular admin levels
}
\author{
Qianyu Dong
}
