% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AN_ANEM_W_ANY.R
\name{AN_ANEM_W_ANY}
\alias{AN_ANEM_W_ANY}
\title{AN_ANEM_W_ANY
IRdata
Percentage of women aged 15-49 classified as having any anemia}
\usage{
AN_ANEM_W_ANY(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
AN_ANEM_W_ANY
IRdata
Percentage of women aged 15-49 classified as having any anemia
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "AN_ANEM_W_ANY",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::AN_ANEM_W_ANY)
}
}
\author{
Qianyu Dong
}
