% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH_DIAT_C_NOT.R
\name{CH_DIAT_C_NOT}
\alias{CH_DIAT_C_NOT}
\title{CH_DIAT_C_NOT}
\usage{
CH_DIAT_C_NOT(KRdata)
}
\arguments{
\item{KRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
CH_DIAT_C_NOT
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "CH_DIAT_C_NOT", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
