% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{ch_novac_either}
\alias{ch_novac_either}
\title{CH_VACS_C_NON
KR
Children with no vaccinations (age 12-23)}
\usage{
ch_novac_either(Rdata)
}
\arguments{
\item{Rdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
CH_VACS_C_NON
KR
Children with no vaccinations (age 12-23)
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "CH_VACS_C_NON",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::ch_novac_either)
}
}
\author{
Qianyu Dong
}
