% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{ch_pent3_either}
\alias{ch_pent3_either}
\title{CH_VACC_C_DP3
DPT3  KR
Percentage of children (age 12-23)
Pentavalent 3rd dose vaccination according to either source"}
\usage{
ch_pent3_either(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
CH_VACC_C_DP3
DPT3  KR
Percentage of children (age 12-23)
Pentavalent 3rd dose vaccination according to either source"
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "CH_VACC_C_DP3",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::ch_pent3_either)
}
}
\author{
Qianyu Dong
}
