% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH_SZWT_C_L25.R
\name{CH_SZWT_C_L25}
\alias{CH_SZWT_C_L25}
\title{CH_SZWT_C_L25}
\usage{
CH_SZWT_C_L25(KRdata)
}
\arguments{
\item{KRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
CH_SZWT_C_L25
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "CH_SZWT_C_L25", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
