% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterinfo.R
\name{clusterinfo}
\alias{clusterinfo}
\title{Get cluster information}
\usage{
clusterinfo(geo, poly.adm1, poly.adm2)
}
\arguments{
\item{geo}{spatial point dataframe}

\item{poly.adm1}{spatial polygons dataframe for admin 1}

\item{poly.adm2}{spatial polygons dataframe for admin 2.}
}
\value{
This function returns the dataset that contains admin 1 and admin 2 information and coordinates for each cluster.
\itemize{
  \item cluster.info
}
}
\description{
This function add admin 1 and admin2 information to a paticular DHS survey.
}
\examples{
\dontrun{
}

}
\author{
Qianyu Dong
}
