% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterModel.R
\name{clusterModel}
\alias{clusterModel}
\title{Calculate cluster model estimates using beta binomial model}
\usage{
clusterModel(
  data,
  cluster.info,
  admin.info,
  admin,
  CI = 0.95,
  model = c("bym2", "iid"),
  stratification = FALSE,
  aggregation = FALSE,
  overdisp.mean = 0,
  overdisp.prec = 0.4
)
}
\arguments{
\item{data}{dataframe that contains the indicator of interests}

\item{cluster.info}{dataframe that contains admin 1 and admin 2 information and coordinates for each cluster.}

\item{admin.info}{dataframe that contains population and urban/rural proportion at specific admin level}

\item{admin}{admin level for the model}

\item{CI}{Credible interval to be used. Default to 0.95.}

\item{model}{smoothing model used in the random effect. Options are independent ("iid") or spatial ("bym2").}

\item{stratification}{whether or not to include urban/rural stratum.}

\item{aggregation}{whether or not report aggregation results.}

\item{overdisp.mean}{prior mean for logit (d).}

\item{overdisp.prec}{prior precision for logit (d).}
}
\value{
This function returns the dataset that contain district name and population for given  tiff files and polygons of admin level,
  \item { clusterModel.result
}
}
\description{
This function calculate smoothed direct estimates at given admin level.
}
\examples{
\dontrun{
}

}
\author{
Qianyu Dong
}
