% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM_ECMR_C_NNF.R
\name{CM_ECMR_C_NNF}
\alias{CM_ECMR_C_NNF}
\title{CM_ECMR_C_NNF NMR five years prior to survey.}
\usage{
CM_ECMR_C_NNF(Rdata)
}
\arguments{
\item{Rdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
BR
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
