% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CN_NUTS_C_WAP.R
\name{CN_NUTS_C_WAP}
\alias{CN_NUTS_C_WAP}
\title{CN_NUTS_C_WAP}
\usage{
CN_NUTS_C_WAP(PRdata)
}
\arguments{
\item{PRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
CN_NUTS_C_WAP
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
