% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CO_MOBB_W_MOB.R
\name{CO_MOBB_W_MOB}
\alias{CO_MOBB_W_MOB}
\title{CO_MOBB_W_MOB}
\usage{
CO_MOBB_W_MOB(IRdata)
}
\arguments{
\item{IRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
CO_MOBB_W_MOB
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "CO_MOBB_W_MOB", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
