% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP_BREG_C_REG.R
\name{CP_BREG_C_REG}
\alias{CP_BREG_C_REG}
\title{CP_BREG_C_REG}
\usage{
CP_BREG_C_REG(PRdata)
}
\arguments{
\item{PRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
CP_BREG_C_REG
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "CP_BREG_C_REG", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
