% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directEST.R
\name{directEST}
\alias{directEST}
\title{Calculate direct estimates}
\usage{
directEST(dat.tem, clusterinfo, admininfo, admin, Amat)
}
\arguments{
\item{dat.tem}{dataframe that contains the indicator of interests}

\item{clusterinfo}{dataframe that contains admin 1 and admin 2 information and coordinates for each cluster.}

\item{admininfo}{dataframe that contains population and urban/rural proportion at specific admin level}

\item{admin}{admin level for the model}
}
\value{
This function returns the dataset that contain district name and population for given  tiff files and polygons of admin level,
  \item { directEST.result
}
}
\description{
This function calculate direct estimates at given admin level.
}
\examples{
\dontrun{
}

}
\author{
Qianyu Dong
}
