% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directEST.R
\name{directEST}
\alias{directEST}
\title{Calculate direct estimates}
\usage{
directEST(
  data,
  cluster.info,
  admin,
  strata = "all",
  CI = 0.95,
  weight = c("population", "survey")[1],
  admin.info = NULL,
  aggregation = FALSE
)
}
\arguments{
\item{data}{dataframe that contains the indicator of interests}

\item{cluster.info}{dataframe that contains admin 1 and admin 2 information and coordinates for each cluster.}

\item{admin}{admin level for the model}

\item{strata}{admin level for the model}

\item{CI}{Credible interval to be used. Default to 0.95.}

\item{weight}{the weight used for aggregating result, "population" or "survey"}

\item{admin.info}{list from the output of adminInfo function}

\item{aggregation}{whether or not report aggregation results.}
}
\value{
This function returns the dataset that contain district name and population for given  tiff files and polygons of admin level,
  \item { directEST.result
}
}
\description{
This function calculate direct estimates at given admin level.
}
\examples{
\dontrun{
}

}
\author{
Qianyu Dong
}
