% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directEST.R
\name{directEST}
\alias{directEST}
\title{Calculate direct estimates}
\usage{
directEST(data, cluster.info, admin.info, admin, strata, weight)
}
\arguments{
\item{data}{dataframe that contains the indicator of interests}

\item{cluster.info}{dataframe that contains admin 1 and admin 2 information and coordinates for each cluster.}

\item{admin.info}{list from the output of adminInfo function}

\item{admin}{admin level for the model}

\item{strata}{admin level for the model}

\item{weight}{the weight used for aggregating result, "population" or "SamplingWeight"}
}
\value{
This function returns the dataset that contain district name and population for given  tiff files and polygons of admin level,
  \item { directEST.result
}
}
\description{
This function calculate direct estimates at given admin level.
}
\examples{
\dontrun{
}

}
\author{
Qianyu Dong
}
