% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV_AFSV_W_A10.R
\name{DV_AFSV_W_A10}
\alias{DV_AFSV_W_A10}
\title{DV_AFSV_W_A10}
\usage{
DV_AFSV_W_A10(IRdata)
}
\arguments{
\item{IRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator

A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
DV_AFSV_W_A10
}
\examples{
\dontrun{
}
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "DV_AFSV_W_A10", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
