% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV_PCPV_W_EMP.R
\name{DV_PCPV_W_EMP}
\alias{DV_PCPV_W_EMP}
\title{DV_PCPV_W_EMP}
\usage{
DV_PCPV_W_EMP(IRdata)
}
\arguments{
\item{IRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
DV_PCPV_W_EMP
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
