% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV_STPS_W_BFR.R
\name{DV_STPS_W_BFR}
\alias{DV_STPS_W_BFR}
\title{DV_STPS_W_BFR}
\usage{
DV_STPS_W_BFR(IRdata)
}
\arguments{
\item{IRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
DV_STPS_W_BFR
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
