% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DV_VPRG_W_VPG.R
\name{DV_VPRG_W_VPG}
\alias{DV_VPRG_W_VPG}
\title{DV_VPRG_W_VPG}
\usage{
DV_VPRG_W_VPG(IRdata)
}
\arguments{
\item{IRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
DV_VPRG_W_VPG
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "DV_VPRG_W_VPG", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
