% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_EMPM_W_EMP.R
\name{EM_EMPM_W_EMP}
\alias{EM_EMPM_W_EMP}
\title{EM_EMPM_W_EMP}
\usage{
EM_EMPM_W_EMP(IRdata)
}
\arguments{
\item{IRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
EM_EMPM_W_EMP
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
