% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_CUSA_W_WTH.R
\name{FP_CUSA_W_WTH}
\alias{FP_CUSA_W_WTH}
\title{FP_CUSA_W_WTH}
\usage{
FP_CUSA_W_WTH(IRdata)
}
\arguments{
\item{IRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
FP_CUSA_W_WTH
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
