% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_EFPM_M_NWS.R
\name{FP_EFPM_M_NWS}
\alias{FP_EFPM_M_NWS}
\title{FP_EFPM_M_NWS}
\usage{
FP_EFPM_M_NWS(MRdata)
}
\arguments{
\item{MRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
FP_EFPM_M_NWS
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
