% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_EFPM_M_TLV.R
\name{FP_EFPM_M_TLV}
\alias{FP_EFPM_M_TLV}
\title{FP_EFPM_M_TLV}
\usage{
FP_EFPM_M_TLV(MRdata)
}
\arguments{
\item{MRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
FP_EFPM_M_TLV
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "FP_EFPM_M_TLV", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
