% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FP_KMTA_W_FST.R
\name{FP_KMTA_W_FST}
\alias{FP_KMTA_W_FST}
\title{FP_KMTA_W_FST}
\usage{
FP_KMTA_W_FST(IRdata)
}
\arguments{
\item{IRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
FP_KMTA_W_FST
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
