% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{fp_unmet_tot}
\alias{fp_unmet_tot}
\title{Internal function to compute "Any anemia - women"}
\usage{
fp_unmet_tot(Rdata)
}
\arguments{
\item{RData}{list returned from getDHSdata when indicator is specified as NULL}
}
\description{
Internal function to compute "Any anemia - women"
}
