% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDHSdata.R
\name{getDHSdata}
\alias{getDHSdata}
\title{Download DHS data}
\usage{
getDHSdata(country, survey)
}
\arguments{
\item{country}{add description to country argument here.}

\item{survey}{add description to survey argument here.}
}
\value{
This function returns a list of datasets containing:
\itemize{
  \item IRdata individual file
  \item PRdata 
  \item KRdata 
}
}
\description{
This function downloads all relevant DHS data for a particular country and survey.
}
\examples{
\dontrun{
}

}
\author{
Qianyu Dong
}
