% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDHSdata.R
\name{getDHSdata}
\alias{getDHSdata}
\title{Download DHS survey data}
\usage{
getDHSdata(country, indicator, year)
}
\arguments{
\item{country}{Country name.}

\item{indicator}{Indicator of interests. Current list of supported indicators include: "womananemia", "ancvisit4+", "stunting", "wasting", "DPT3".}

\item{year}{Year the survey conducted.}
}
\value{
This function returns the survey dataset that contains the indicator.
}
\description{
This function downloads DHS data for a particular country and survey.
}
\examples{
\dontrun{
dhsData <-surveyPrev:: getDHSdata(country = "Zambia", 
                                 indicator = "ancvisit4+", 
                                 year = 2018)
}

}
\author{
Qianyu Dong
}
