% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDHSdata.R
\name{getDHSdata}
\alias{getDHSdata}
\title{Download DHS survey data}
\usage{
getDHSdata(country, indicator, year)
}
\arguments{
\item{country}{Country name.}

\item{indicator}{Indicator of interests.}

\item{year}{Year the survey conducted.}
}
\value{
This function returns the survey dataset that contains the indicator:
\itemize{
  \item Rdata
}
}
\description{
This function downloads DHS data for a particular country and survey.
}
\examples{
\dontrun{
}

}
\author{
Qianyu Dong
}
