% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDHSindicator.R
\name{getDHSindicator}
\alias{getDHSindicator}
\title{Process DHS data}
\usage{
getDHSindicator(Rdata, indicator)
}
\arguments{
\item{Rdata}{Result from getDHSdata function, the raw DHS survry data from get_datasets.}

\item{indicator}{Indicator of interests.}
}
\value{
The function returns processed survey data that contains the indicator of interests.
}
\description{
This function processes DHS data from getDHSdata function.
}
\examples{
\dontrun{
dhsData <-surveyPrev::getDHSdata(country = "Zambia", 
                                 indicator = "ancvisit4+", 
                                 year = 2018)

data <- getDHSindicator(dhsData, indicator = "ancvisit4+")
}

}
\author{
Qianyu Dong
}
