% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HA_AFSY_W_A18.R
\name{HA_AFSY_W_A18}
\alias{HA_AFSY_W_A18}
\title{HA_AFSY_W_A18}
\usage{
HA_AFSY_W_A18(IRdata)
}
\arguments{
\item{IRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
HA_AFSY_W_A18
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "HA_AFSY_W_A18", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
