% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HA_CATH_W_ATN.R
\name{HA_CATH_W_ATN}
\alias{HA_CATH_W_ATN}
\title{HA_CATH_W_ATN}
\usage{
HA_CATH_W_ATN(IRdata)
}
\arguments{
\item{IRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator

A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator

A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
HA_CATH_W_ATN
}
\examples{
\dontrun{
}
\dontrun{
}
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "HA_CATH_W_ATN", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
