% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HA_HRSX_W_CND.R
\name{HA_HRSX_W_CND}
\alias{HA_HRSX_W_CND}
\title{HA_HRSX_W_CND}
\usage{
HA_HRSX_W_CND(IRdata)
}
\arguments{
\item{IRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
HA_HRSX_W_CND
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "HA_HRSX_W_CND", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
