% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HA_HVST_M_HRD.R
\name{HA_HVST_M_HRD}
\alias{HA_HVST_M_HRD}
\title{HA_HVST_M_HRD}
\usage{
HA_HVST_M_HRD(MRdata)
}
\arguments{
\item{MRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
HA_HVST_M_HRD
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
