% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HA_HVTY_M_TRR.R
\name{HA_HVTY_M_TRR}
\alias{HA_HVTY_M_TRR}
\title{HA_HVTY_M_TRR}
\usage{
HA_HVTY_M_TRR(MRdata)
}
\arguments{
\item{MRdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
HA_HVTY_M_TRR
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "HA_HVTY_M_TRR", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
