% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HA_KAID_W_HRD.R
\name{HA_KAID_W_HRD}
\alias{HA_KAID_W_HRD}
\title{HA_KAID_W_HRD Percentage of women who have heard of HIV or AIDS
hk_ever_heard in github
IR}
\usage{
HA_KAID_W_HRD(Rdata)
}
\arguments{
\item{Rdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
HA_KAID_W_HRD Percentage of women who have heard of HIV or AIDS
hk_ever_heard in github
IR
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
