% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HA_STIS_M_SOR.R
\name{HA_STIS_M_SOR}
\alias{HA_STIS_M_SOR}
\title{HA_STIS_M_SOR}
\usage{
HA_STIS_M_SOR(MRdata)
}
\arguments{
\item{MRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
HA_STIS_M_SOR
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
