% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HA_STIS_W_STI.R
\name{HA_STIS_W_STI}
\alias{HA_STIS_W_STI}
\title{HA_STIS_W_STI  Percentage of women reporting a sexually transmitted infection in the 12 months preceding the survey among women who ever had sexual intercourse
hk_sti in github
IR}
\usage{
HA_STIS_W_STI(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
HA_STIS_W_STI  Percentage of women reporting a sexually transmitted infection in the 12 months preceding the survey among women who ever had sexual intercourse
hk_sti in github
IR
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
