% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HC_FLRM_H_CER.R
\name{HC_FLRM_H_CER}
\alias{HC_FLRM_H_CER}
\title{HC_FLRM_H_CER Percentage of households with ceramic tile floors
ph_floor in github
HR}
\usage{
HC_FLRM_H_CER(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
HC_FLRM_H_CER Percentage of households with ceramic tile floors
ph_floor in github
HR
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
