% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervalPlot.R
\name{intervalPlot}
\alias{intervalPlot}
\title{Get scatter plot for any two model results}
\usage{
intervalPlot(res)
}
\arguments{
\item{res}{model result  using surveyPrev}

\item{line1}{yintercept}
}
\value{
This function returns the dataset that contain district name and population for given  tiff files and polygons of admin level
  \item {plot
}
}
\description{
This function return scatter plot at admin 1 level for any two model results
}
\examples{
\dontrun{
}

}
\author{
Qianyu Dong
}
