% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MA_CWIV_W_CWV.R
\name{MA_CWIV_W_CWV}
\alias{MA_CWIV_W_CWV}
\title{MA_CWIV_W_CWV}
\usage{
MA_CWIV_W_CWV(IRdata)
}
\arguments{
\item{IRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
MA_CWIV_W_CWV
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
