% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{nt_ebf}
\alias{nt_ebf}
\title{CN_BRFS_C_EXB
Children exclusively breastfed
NT_IYCF.do KR "Exclusively breastfed - last-born under 6 months"
Children exclusively breastfed (Prevalence of exclusive breastfeeding of children under six months of age)}
\usage{
nt_ebf(Rdata)
}
\arguments{
\item{Rdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in  survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
CN_BRFS_C_EXB
Children exclusively breastfed
NT_IYCF.do KR "Exclusively breastfed - last-born under 6 months"
Children exclusively breastfed (Prevalence of exclusive breastfeeding of children under six months of age)
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia",
                                 indicator = "CN_BRFS_C_EXB",
                                 year = 2018)
data <- getDHSindicator(dhsData, indicator = NULL,
                         FUN = surveyPrev::nt_ebf)
}
}
\author{
Qianyu Dong
}
