% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PR_DESL_W_WNM.R
\name{PR_DESL_W_WNM}
\alias{PR_DESL_W_WNM}
\title{PR_DESL_W_WNM}
\usage{
PR_DESL_W_WNM(IRdata)
}
\arguments{
\item{IRdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
PR_DESL_W_WNM
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
