% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RH_DELP_C_DHF.R
\name{RH_DELP_C_DHF}
\alias{RH_DELP_C_DHF}
\title{RH_DELP_C_DHF}
\usage{
RH_DELP_C_DHF(Rdata)
}
\arguments{
\item{Rdata}{data.frame from survryPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in survryPrev::getDHSindicator. The whole function can be used as a parameter in survryPrev::getDHSindicator
}
\description{
RH_DELP_C_DHF
}
\examples{
\dontrun{
}
}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
