% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlot.R
\name{scatterPlot}
\alias{scatterPlot}
\title{Get scatter plot for any two model results}
\usage{
scatterPlot(scatter, res1, nameres1, res2, nameres2)
}
\arguments{
\item{res1}{model result 1 using surveyPrev}

\item{nameres1}{name of model 1}

\item{res2}{model result 2 using surveyPrev}

\item{nameres2}{name of model 2}
}
\value{
This function returns the dataset that contain district name and population for given  tiff files and polygons of admin level
  \item {plot
}
}
\description{
This function return scatter plot at admin 1 level for any two model results
}
\examples{
\dontrun{
}

}
\author{
Qianyu Dong
}
