% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WS_SRCE_H_USG.R
\name{WS_SRCE_H_USG}
\alias{WS_SRCE_H_USG}
\title{WS_SRCE_H_USG Percentage of households whose main source of drinking water is an unprotected spring
ph_wtr_source in github "unprotected spring" = 42,
HR}
\usage{
WS_SRCE_H_USG(Rdata)
}
\arguments{
\item{Rdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
WS_SRCE_H_USG Percentage of households whose main source of drinking water is an unprotected spring
ph_wtr_source in github "unprotected spring" = 42,
HR
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "WS_SRCE_H_USG", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
