% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WS_WTRT_P_STN.R
\name{WS_WTRT_P_STN}
\alias{WS_WTRT_P_STN}
\title{WS_WTRT_P_STN}
\usage{
WS_WTRT_P_STN(WASHdata)
}
\arguments{
\item{WASHdata}{data.frame from surveyPrev::getDHSdata}
}
\value{
A partially processed data.frame that will be used in surveyPrev::getDHSindicator. The whole function can be used as a parameter in surveyPrev::getDHSindicator
}
\description{
WS_WTRT_P_STN
}
\examples{
\dontrun{
dhsData <- getDHSdata(country = "Zambia", indicator = "WS_WTRT_P_STN", year = 2018)
}

}
\author{
Miaolei Bao, Yunhan Wu, Qianyu Dong
}
